requirements.txt
fastapi
uvicorn
streamlit
pandas
polars
duckdb
psycopg2-binary
sqlalchemy
alembic
lightgbm
mlflow==2.9.2
redis
rq
scikit-learn
shap
matplotlib
plotly
python-dotenv
httpx
pydantic
pydantic-settings
watchdog
python-multipart
pyarrow
category_encoders
featuretools
optuna